+++ CHANGELOG.md
@@ -4,6 +4,26 @@
 
 [1]: https://www.npmjs.com/package/gax-nodejs?activeTab=versions
 
+## [2.12.0](https://www.github.com/googleapis/gax-nodejs/compare/v2.11.3...v2.12.0) (2021-04-29)
+
+
+### Features
+
+* GrpcClient.loadProtoJSON to load protobuf.js JSON proto ([#985](https://www.github.com/googleapis/gax-nodejs/issues/985)) ([819b447](https://www.github.com/googleapis/gax-nodejs/commit/819b447964a635ef2ac75769c0f6eadbd202ddfa))
+
+
+### Bug Fixes
+
+* **deps:** update dependency @grpc/grpc-js to ~1.3.0 ([#990](https://www.github.com/googleapis/gax-nodejs/issues/990)) ([f131ee0](https://www.github.com/googleapis/gax-nodejs/commit/f131ee0fb0411f57f43777ca7e3dcec6556115f2))
+
+### [2.11.3](https://www.github.com/googleapis/gax-nodejs/compare/v2.11.2...v2.11.3) (2021-04-22)
+
+
+### Bug Fixes
+
+* fallback can now be a string ([#983](https://www.github.com/googleapis/gax-nodejs/issues/983)) ([d4e495a](https://www.github.com/googleapis/gax-nodejs/commit/d4e495af46176b8dd74595108f25a6df99ee042d))
+* imports for node-fetch when using webpack ([#972](https://www.github.com/googleapis/gax-nodejs/issues/972)) ([505c6b8](https://www.github.com/googleapis/gax-nodejs/commit/505c6b8a3a53f1875617a2258869e4646a0e439b))
+
 ### [2.11.2](https://www.github.com/googleapis/gax-nodejs/compare/v2.11.1...v2.11.2) (2021-03-11)